refactor: migrate from migrations/ to db/migrations/
The old migrations/ directory contained SQLite-only migrations that
have been superseded by the dual-database structure in db/migrations/.

Changes:
- Update all test file imports from "migrations" to "db/migrations"
- Remove legacy migrations/ directory (15 SQL files + metadata)
- All migrations now live in db/migrations/{sqlite,postgres}/

The db/migrations/ structure supports both SQLite and PostgreSQL
with parallel migration files, while the old migrations/ was
SQLite-only and no longer referenced by production code.

Files updated:
- internal/app_test.go
- internal/base/base_test.go
- internal/handlers/private/install_test.go
- internal/mailer/mailer_test.go
- internal/queries/queries_test.go
- internal/queries/setting_test.go
- internal/routes/api_routes_test.go
- internal/testutil/testdb.go
- internal/webhook/hook_test.go
