+++ internal/app_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/gofiber/fiber/v3"
 
 	"github.com/shurco/mycart/internal/queries"
-	"github.com/shurco/mycart/migrations"
+	"github.com/shurco/mycart/db/migrations"
 	"github.com/shurco/mycart/pkg/logging"
 )
