+++ internal/webhook/hook_test.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/shurco/mycart/internal/models"
 	"github.com/shurco/mycart/internal/queries"
-	"github.com/shurco/mycart/migrations"
+	"github.com/shurco/mycart/db/migrations"
 	"github.com/shurco/mycart/pkg/litepay"
 )
